refactor: move cursor/orderBy utils to graphql/utils, use isPlainObject typeguard
