+++ packages/twenty-front/src/modules/object-record/graphql/utils/isOrderByDirection.ts
