+++ packages/twenty-front/src/modules/object-record/record-show/hooks/useRecordShowPagePagination.ts
@@ -4,9 +4,9 @@ import { useParams, useSearchParams } from 'react-router-dom';
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
 import { useFindManyRecords } from '@/object-record/hooks/useFindManyRecords';
 import { lastShowPageRecordIdState } from '@/object-record/record-field/ui/states/lastShowPageRecordId';
-import { computeCursorArgFilter } from '@/object-record/record-show/utils/computeCursorArgFilter';
-import { extractOrderByFieldNames } from '@/object-record/record-show/utils/extractOrderByFieldNames';
-import { reverseOrderBy } from '@/object-record/record-show/utils/reverseOrderBy';
+import { computeCursorArgFilter } from '@/object-record/graphql/utils/computeCursorArgFilter';
+import { extractOrderByFieldNames } from '@/object-record/graphql/utils/extractOrderByFieldNames';
+import { reverseOrderBy } from '@/object-record/graphql/utils/reverseOrderBy';
 import { useSetAtomState } from '@/ui/utilities/state/jotai/hooks/useSetAtomState';
 import { useQueryVariablesFromParentView } from '@/views/hooks/useQueryVariablesFromParentView';
 import { AppPath } from 'twenty-shared/types';
